requirements.txt
playwright
python-dotenv
pyyaml
